requirements.txt
Flask==3.1.0
pika==1.3.2
psycopg2-binary==2.9.10
